Adding cargo-tracker-infrastructure-axon module.
This module contains Axon infrastructure that is not generic enough to be used as general purpose library, but rather relies on some other modules already present in the project, namely 'api-open-shared' module.
In the architecture it logically is above general purpose infrastructure projects, but bellow subdomain infrastructure projects or applications.