Changing yml configuration files into properties configuration files.
Main reason for this change is inadequate tooling support for yaml in IDEA. IDEA constantly insists on breaking down yaml entries into individual sorted components.
This is a problem for custom ordering of properties and maintaining desired order of properties in the configuration. Properties parts are much harder to find with
a text search then when whole property is in a single line. Besides, at the moment I do not really need any advanced feature of yaml.