Replacing groovy assert with assertj to improve code coverage.
Downside is necessity to have assertj library (4.8 M) in every module, even those at the language level.
Another downside (actually this is the same as with groovy assert) that programming by contracts is not clearly expressed. Everything is assert without additional semantic meaning.
This also includes throwing AssertionError instead of more expressive IllegalArgumentException for preconditions or IllegalStateException for postconditions.