test: Refactor integration tests with varying configurations of axon gateways
When having multiple application instances with a single Axon Server, creation commands
can be sent to any instance, even if they have different configurations (which is the
case for some of our integration tests).

As there is no way to force the Axon Server to send creation commands to a particular
application instance, integration tests with different configurations are refactored
to use different Axon Server instances.

This is a better solution than using DirtiesContext annotations. We are paying the
price only once at the static initialization of all tests instead of recreating the
application context for each test class. This is especially important as the number
of integration tests for applications can grow significantly.