feature: Add cargotracker-lib-boundary-query-api module
This module is needed as a place for hosting boundary classes specific only for query applications.
Currently, those are DTOs required for paging and sorting query results.

We could add those classes to the existing cargotracker-lib-boundary-api. However, the classes in
question are required only for query applications, and those classes require a new dependency on
the validation library. And those are valid reasons for introducing a new module.