Isolate Groovy version and libs for Micronaut
At this point (3.7.2 version), Micronaut uses Groovy 3 as a dependency. Therefore we are isolating Groovy 3 and related
dependencies specifically for Micronaut.

The next Major version of Micronut will use Groovy 4. However, we will not unify Micronaut's Groovy version with the
rest to be prepared for Groovy version mismatches in the future.