fix: Fix a test caused by difference in Instant precision between PostgreSQL (micro) and Linux (nano)
Default and maximum precision of timestamptz in PostgreSLQ is microsecond. Since Instant precision in
Java 17 is increased (from Java 11) to nanoseconds, for instant comparison we have to truncate
precision on milliseconds. One would probably expect that we should truncate to microseconds, but
timestamps stored in PostgreSQL are rounded at the microsecond and comparison with truncation can fail.
We can either implement rounding for Java instant, or just use truncation with the millis precision.