Decouple event structure from value objects structure
Primary intention is to have a serialization format of events which is largely
independent of internal structure of value objects.

Unfortunatelly, this commit also includes several additional refactorings like
the change from a single commodity to the collection of commodities in event, and
renaming of event properties:
commodity -> commodities
bookingTotalCommodityWeight -> commodityTotalWeight,
bookingTotalContainerTeuCount -> commodityTotalContainerTeuCount

Also note that this is initial commit of solution. It is still required to do a
series of smaller refactorings to accommodate mention changes accross the codebase.