test: Upgrade to PostgreSQL 15.2 and update tests and db init scripts
PostgreSQL's initialization script have to be updated to add all privileges on public schema for db_migration user.
This is due to change in granting privileges in PostgreSQL 15 and above
REFERENCE - https://www.cybertec-postgresql.com/en/error-permission-denied-schema-public

NOTE - there will be a flyway warning about upgrading to a version that supports PostgreSQL 15. This will be resolved
with spring boot 3.x upgrade which updates flyway version. Beside the warning, everything works as expected.