deps: Upgrade Micronaut to 4.2.2 version
This is a significant upgrade because of Micronaut switch from Groovy 3 to 4.
That causes a bunch of changes in dependencies, but the most important, changes
to the GraalVM native image build.

Those changes can be observed in klokwrk-tool-gradle-source-repack/build.gradle.
The most challenging one was to discover the need for reconfiguring
groovy.indy.optimize.threshold and groovy.indy.fallback.threshold system properties.